+++ Assets/Packages/DraasGames.Core/Runtime/UI/Extensions/Editor/CustomSelectableOdinEditor.cs
@@ -1,0 +1,12 @@
+#if UNITY_EDITOR
+using Sirenix.OdinInspector.Editor;
+using UnityEditor;
+
+namespace DraasGames.Core.Runtime.UI.Extensions.Editor
+{
+    [CustomEditor(typeof(CustomSelectable))]
+    public class CustomSelectableOdinEditor : OdinEditor
+    {
+    }
+}
+#endif